+++ node/cli/src/chain_spec/testnet.rs
@@ -55,6 +55,8 @@ type TssEndpoint = String;
 /// Custom input data for building the chainspec for a particular test network
 #[derive(Serialize, Deserialize, Clone, Debug, Default)]
 pub struct TestnetChainSpecInputs {
+    /// Initial authorities - account IDs for the initial validators
+    pub initial_authorities: Option<Vec<InitialAuthority>>,
     /// A map of hostname / socket address to [TssPublicKeys] of the TSS servers
     /// [TssPublicKeys] is the output type returned from the TSS server `/info` http route
     pub tss_details: HashMap<String, TssPublicKeys>,
@@ -84,6 +86,24 @@ impl TestnetChainSpecInputs {
         let input = std::fs::read(path).map_err(|e| format!("{e:?}"))?;
         serde_json::from_slice(&input).map_err(|e| format!("{e:?}"))
     }
+}
+
+/// Public account IDs for an initial validator
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct InitialAuthority {
+    /// The stash account ID. This holds staked funds
+    pub stash: AccountId,
+    /// The controller account ID. This is used for sending staking-related extrinsics, allowing the
+    /// stash account can be kept 'cold'
+    pub controller: AccountId,
+    /// GRANDPA account ID (Ed25519) - Session key for finality
+    pub grandpa: GrandpaId,
+    /// BABE account ID - Session key for the block production mechanism
+    pub babe: BabeId,
+    /// Im online account ID - Session key for tracking responsiveness
+    pub im_online: ImOnlineId,
+    /// Authority discovery account - Session key for finding other validators
+    pub authority_discovery: AuthorityDiscoveryId,
 }
 
 pub fn testnet_local_initial_authorities(
@@ -244,6 +264,31 @@ pub fn testnet_local_initial_tss_servers(
 ///  - Update all the accounts here using keys you control, or
 ///  - Run the `testnet-local` config, which uses well-known keys
 pub fn testnet_config(inputs: TestnetChainSpecInputs) -> Result<ChainSpec, String> {
+    let initial_authorities = match inputs.initial_authorities {
+        Some(authorities) => {
+            let mut initial_authorities: Vec<(
+                AccountId,
+                AccountId,
+                GrandpaId,
+                BabeId,
+                ImOnlineId,
+                AuthorityDiscoveryId,
+            )> = vec![];
+            for initial_authority in authorities {
+                initial_authorities.push((
+                    initial_authority.stash,
+                    initial_authority.controller,
+                    initial_authority.grandpa,
+                    initial_authority.babe,
+                    initial_authority.im_online,
+                    initial_authority.authority_discovery,
+                ))
+            }
+            initial_authorities
+        },
+        None => testnet_initial_authorities(),
+    };
+
     let tss_details = inputs
         .tss_details
         .into_iter()
@@ -273,7 +318,7 @@ pub fn testnet_config(inputs: TestnetChainSpecInputs) -> Result<ChainSpec, Strin
         .with_id("entropy_testnet")
         .with_chain_type(ChainType::Live)
         .with_genesis_config_patch(testnet_genesis_config(
-            testnet_initial_authorities(),
+            initial_authorities,
             vec![],
             hex!["b848e84ef81dfeabef80caed10d7d34cc10e98e71fd00c5777b81177a510d871"].into(),
             tss_details,