+++ CHANGELOG.md
@@ -74,6 +74,7 @@ runtime
 - Database encryption key backup / recovery feature for entropy-tss [(#1249)](https://github.com/entropyxyz/entropy-core/pull/1249)
 - Add no hash option [(#1266)](https://github.com/entropyxyz/entropy-core/pull/1266)
 - Template testnet chainspec artifact ([#1442](https://github.com/entropyxyz/entropy-core/pull/1442))
+- Add initial authorities to testnet chainspec inputs [(#1451)](https://github.com/entropyxyz/entropy-core/pull/1451)
 
 ### Changed
 - Use correct key rotation endpoint in OCW ([#1104](https://github.com/entropyxyz/entropy-core/pull/1104))